Develop (#139)
* Moving breakout code to closer to CreateProcess call (#129)

* Changing sing location for Psfmonitor

* Adding output for PsfMonitor to PsfMonitor directory

* Moving PsfMonitor stuff to it's own folder

* Adding new files ot nuspec

* Changing sing location for Psfmonitor

* Adding output for PsfMonitor to PsfMonitor directory

* Moving PsfMonitor stuff to it's own folder

* Fixing issues

* Using correct WIL macro

* Moving code so atrribute list is in scopr to CreateProcess


* Full PrivateProfile support (#124)

* Complete PrivateProfile intercepts for FRF

* Full PrivateProfile support

* Full PrivateProfile support

* Added PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* Various fixes for application compatibility (#132)

* Complete PrivateProfile intercepts for FRF

* Full PrivateProfile support

* Full PrivateProfile support

* Added PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* Replacement Post350 (#137)

* Complete PrivateProfile intercepts for FRF

* Full PrivateProfile support

* Full PrivateProfile support

* Added PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* PrivateProfile Tests

* Changes for script flexibility: Can be in any folder inside or outside package, may use quotation marks, may use new pseudovariables to reference the package root or user redirection area.

* Updates to documentation on scripting.

* Redo of Post350 Changes

* Update for naming conventions

* Added optional executionpolicy options to script configs.

* Correction to scripting Bypass

* Scripting documentation for ExecutionPolicy

* Fix issue with sending files to StartingScriptWrapper.

* Put wait back in for Shell Launch

* Requested cleanups to PR137

* Document UrlDecode method.

* Cleanup StripFileColonSlash

* wrong merge fix for calling StartProcess


* Target and Script pseudo-variables and cleaup of PR137 manual merge issues. (#138)

* Making sure breakout does not occur with PS scripts

+++ fixups/FileRedirectionFixup/DeleteFileFixup.cpp
@@ -14,28 +14,36 @@ BOOL __stdcall DeleteFileFixup(_In_ const CharT* fileName) noexcept
     {
         if (guard)
         {
-            LogString(L"DeleteFileFixup for fileName", fileName);
+            DWORD DeleteFileInstance = ++g_FileIntceptInstance;
+            LogString(DeleteFileInstance,L"DeleteFileFixup for fileName", fileName);
             
-
-            // NOTE: This will only delete the redirected file. If the file previously existed in the package path, then
-            //       it will remain there and a later attempt to open, etc. the file will succeed. In the future, if
-            //       this proves to be an issue, we could maintain a collection of package files that have been
-            //       "deleted" and then just pretend like they've been deleted. Such a change would be rather large and
-            //       disruptful and probably fairly inefficient as it would impact virtually every code path, so we'll
-            //       put it off for now.
-            auto [shouldRedirect, redirectPath, shoudReadonly] = ShouldRedirect(fileName, redirect_flags::none);
-            if (shouldRedirect)
+            if (!IsUnderUserAppDataLocalPackages(fileName))
             {
-                if (!impl::PathExists(redirectPath.c_str()) && impl::PathExists(fileName))
+
+                // NOTE: This will only delete the redirected file. If the file previously existed in the package path, then
+                //       it will remain there and a later attempt to open, etc. the file will succeed. In the future, if
+                //       this proves to be an issue, we could maintain a collection of package files that have been
+                //       "deleted" and then just pretend like they've been deleted. Such a change would be rather large and
+                //       disruptful and probably fairly inefficient as it would impact virtually every code path, so we'll
+                //       put it off for now.
+                auto [shouldRedirect, redirectPath, shoudReadonly] = ShouldRedirect(fileName, redirect_flags::none);
+                if (shouldRedirect)
                 {
-                    // If the file does not exist in the redirected location, but does in the non-redirected location,
-                    // then we want to give the "illusion" that the delete succeeded
-                    return TRUE;
+                    if (!impl::PathExists(redirectPath.c_str()) && impl::PathExists(fileName))
+                    {
+                        // If the file does not exist in the redirected location, but does in the non-redirected location,
+                        // then we want to give the "illusion" that the delete succeeded
+                        return TRUE;
+                    }
+                    else
+                    {
+                        return impl::DeleteFile(redirectPath.c_str());
+                    }
                 }
-                else
-                {
-                    return impl::DeleteFile(redirectPath.c_str());
-                }
+            }
+            else
+            {
+                Log(L"[%d]Under LocalAppData\\Packages, don't redirect", DeleteFileInstance);
             }
         }
     }
